# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for tracking League of Legends summoner spell cooldowns.\nRooms are synchronized with the League Client and updated in real time.\nPostgreSQL is used for persistence and Redis for caching.",
        "title": "LoL Timer API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/current-room": {
            "get": {
                "description": "Returns the room currently synchronized with the League Client.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Get current room",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CurrentRoomResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns OK when the application process is running.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.HealthResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ready": {
            "get": {
                "description": "Checks whether PostgreSQL and Redis are available.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Readiness check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReadinessResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handlers.ReadinessResponse"
                        }
                    }
                }
            }
        },
        "/rooms/{roomId}": {
            "get": {
                "description": "Returns a room with its players and summoner spell cooldowns.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Get room",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room ID",
                        "name": "roomId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Room"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/rooms/{roomId}/players": {
            "post": {
                "description": "Adds a player to an existing room.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Add player",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room ID",
                        "name": "roomId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Player data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePlayerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/rooms/{roomId}/spells/toggle": {
            "post": {
                "description": "Starts a cooldown for a ready spell or resets an active cooldown.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Toggle summoner spell",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room ID",
                        "name": "roomId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Spell toggle data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ToggleSpellRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreatePlayerRequest": {
            "type": "object",
            "properties": {
                "champion": {
                    "type": "string",
                    "example": "Ahri"
                }
            }
        },
        "dto.CurrentRoomResponse": {
            "type": "object",
            "properties": {
                "roomId": {
                    "type": "string",
                    "example": "7936545340-100"
                }
            }
        },
        "dto.ToggleSpellRequest": {
            "type": "object",
            "properties": {
                "gameName": {
                    "type": "string",
                    "example": "PlayerOne"
                },
                "spell": {
                    "type": "string",
                    "example": "Flash"
                },
                "tagLine": {
                    "type": "string",
                    "example": "EUW"
                }
            }
        },
        "handlers.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "handlers.ReadinessResponse": {
            "type": "object",
            "properties": {
                "postgres": {
                    "type": "string",
                    "example": "ok"
                },
                "redis": {
                    "type": "string",
                    "example": "ok"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "models.Player": {
            "type": "object",
            "properties": {
                "champion": {
                    "type": "string"
                },
                "championId": {
                    "type": "integer"
                },
                "gameName": {
                    "type": "string"
                },
                "spells": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SummonerSpell"
                    }
                },
                "summonerSpellHaste": {
                    "type": "integer"
                },
                "tagLine": {
                    "type": "string"
                }
            }
        },
        "models.Room": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "lastUpdated": {
                    "type": "string"
                },
                "players": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Player"
                    }
                }
            }
        },
        "models.SummonerSpell": {
            "type": "object",
            "properties": {
                "baseCooldown": {
                    "type": "integer"
                },
                "cooldownEndTime": {
                    "type": "string"
                },
                "isReady": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "remainingCooldown": {
                    "type": "integer"
                }
            }
        }
    }
}